Move AttrsSensors and VisFlagsWeights
The `data_sources` module seems a closer fit as they are subordinates
of DataSource (and not used anywhere else anyway).
